// src/app/shared/models/service.ts
export class Services {
    id: number;
    name: string;
    displayName: string;
    serviceTypeId: number;
    description: string;
    iconUrl: string;
    colorCode: string;
    isActive: boolean;
    isDelete: boolean;
    createdDate: Date;
    modifiedDate: Date;
    createdBy: number;
    modifiedBy: number;
    serviceTypeName: string;
    constructor() { }
}